Update to allow message and title
